Added profile and delete functionality

+++ techmash/views.py
@@ -23,9 +23,12 @@ from math import fabs
 
 TECHMASH_URL = 'http://www.shaastra.org/2011/media/techmash/'
 @needs_authentication  
-def profile(request):
+def profile(request,username=None):
+    if username == None:
+        username = request.user.username
+        show_buttons = True    
     try:
-        image_list = Photo.objects.filter(user = request.user).order_by('rating')
+        image_list = Photo.objects.filter(user = username).order_by('rating')
     except:
         image_list =list()    
     return render_to_response("techmash/techslam.html", locals(),context_instance= global_context(request))
@@ -106,9 +109,17 @@ def mashphotos(request):
         photo1,photo2=selectimages()
         return render_to_response("techmash/compare.html", locals(),context_instance= global_context(request))
 
-def deleteimage(request):
-    photo_to_delete = Photo.objects.get(id=request.POST['photo_id'])
-    photo_to_delete.delete()
+def deleteimage(request,image_id=None):
+    try:
+        photo_to_delete = Photo.objects.get(id=image_id)
+        if request.user.username == photo_to_delete.user:
+            try:
+                photo_to_delete.delete()
+            except:    
+        else:
+            pass
+    except:
+        pass            
     return HttpResponseRedirect('%stechmash/profile'%settings.SITE_URL)
     
 def selectimages():
